1.3
1.3

+++ 1.3.py
@@ -1,5 +1,5 @@
 print("Введите три числа")
-x = int(input())
-y = int(input())
-z = int(input())
+x = float(input())
+y = float(input())
+z = float(input())
 print((x ** 2 + y ** 2 + z ** 2) ** 0.5)
